separate idea of expression from operand and make MergeMaps exported
